Split the test helpers into their own files
Create two new test modules, `mod common`, which will hold adapter function
to hold common patterns to call `lal::x()` functions used by the test cases.

The test cases themselves go into `mod cases`, and will be split up to
cover high-level features and their typical usage.

With rust tests, adding other files to `tests/` creates extra test
binaries. which we don't want. Instead, `#[test]` cases can be split
between multiple files if they are in a sub modules.
